Enhance Image Pipelines and Documentation Updates (#3)
* Update .gitignore to include .idea directory

Added the .idea directory to the .gitignore file to prevent IDE-specific files from being tracked in the repository. This helps keep the repository clean and focused on project source files.

* Add height and width parameters to image pipelines

Enhanced the DevImg2ImgPipeline and SchnellImg2ImgPipeline classes
by adding height and width parameters. This allows for better control
over the dimensions of the output images during processing.

* Add output format option and update documentation

- Introduced a new command-line argument `--output_format` for specifying the image format of the output (`webp`, `png`, or `jpg`).
- Updated the README.md to include instructions for activating the virtual environment and the new output format option.
- Modified the image saving logic in base_pipeline.py to reflect the selected output format.
- Set the default output format in config.yaml to `webp`.